requirements.txt
Pillow
pyyaml
pynput